Changed casting to size_t

+++ shmht.c
@@ -119,7 +119,7 @@ static PyObject * shmht_open(PyObject *self, PyObject *args)
 
     if (force_init == 0) { //try to load from existing shmht
         mem_size = sizeof(hashtable);
-        if ((long unsigned int)buf.st_size >= sizeof(hashtable)) { //may be valid
+        if ((size_t)buf.st_size >= sizeof(hashtable)) { //may be valid
             ht = mmap(NULL, sizeof(hashtable), PROT_READ|PROT_WRITE, MAP_SHARED, fd, 0);
             if (ht == MAP_FAILED) {
                 PyErr_Format(shmht_error, "mmap failed, map_size=sizeof(hashtable)=%lu: [%d] %s",
@@ -146,7 +146,7 @@ static PyObject * shmht_open(PyObject *self, PyObject *args)
 
     mem_size = ht_memory_size(capacity);
 
-    if ((long unsigned int)buf.st_size < mem_size) {
+    if ((size_t)buf.st_size < mem_size) {
         if (lseek64(fd, mem_size - 1, SEEK_SET) == -1) {
             PyErr_Format(shmht_error, "lseek64 failed: [%d] %s", errno, strerror(errno));
             goto create_failed;
